+++ resources/views/tasks/show.blade.php
@@ -3,6 +3,7 @@
 @section('content')
 
 <!-- ここにページ毎のコンテンツを書く -->
+
  <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
 
     <table class="table table-bordered">
@@ -24,9 +25,10 @@
     {!! link_to_route('tasks.edit', 'このタスクを編集', ['task' => $task->id], ['class' => 'btn btn-light']) !!}
 
 {{-- タスク削除フォーム --}}
+@if (Auth::id() == $task->user_id)
     {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
         {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
     {!! Form::close() !!}
-
+@endif
 
 @endsection
